+++ README.md
@@ -1,2 +1,10 @@
 # seohyeonee-front
 서혀니 웹 프론트
+
+## NPM Script
+`start`: 배포용 빌드 watch
+`dev`: FE 개발 로컬 서버 실행(`webpack-dev-server`)
+`build`: 배포용 빌드 실행
+`lint`: `lint`
+`lint:fix`: `lint --fix`
+`test`: jest 테스트 실행
