+++ package.json
@@ -4,8 +4,9 @@
   "description": "서혀니 웹 프론트",
   "main": "index.js",
   "scripts": {
-    "start": "webpack --mode=production",
-    "dev": "webpack-dev-server",
+    "start": "webpack --config webpack.build.config.js --watch",
+    "dev": "webpack-dev-server --config webpack.development.config.js",
+    "build": "webpack --config webpack.build.config.js",
     "lint": "eslint \"**/*.js\"",
     "lint:fix": "eslint \"**/*.js\" --fix",
     "test": "jest"
@@ -56,7 +57,7 @@
     "react-hot-loader": "^4.12.18",
     "sass-loader": "^8.0.0",
     "style-loader": "^1.0.0",
-    "webpack": "^4.40.2",
+    "webpack": "^4.41.5",
     "webpack-cli": "^3.3.9",
     "webpack-dev-server": "^3.8.1"
   }
